feat(setup_executor.py): remove method `update_inner_referenced_feature_instances()` from variation_executor and move similar method in setup executor
